requirements.txt
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.5
distlib==0.3.6
et-xmlfile==1.1.0
filelock==3.12.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
grimp==2.4
httplib2==0.22.0
identify==2.5.24
idna==3.4
import-linter==1.8.0
iniconfig==2.0.0
lml==0.1.0
mypy==1.2.0
mypy-extensions==1.0.0
nodeenv==1.7.0
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pluggy==1.0.0
pre-commit==3.3.1
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
pyparsing==3.0.9
pytest==7.3.1
PyYAML==6.0
replace_me @ file:///Users/albert/Projects/Pynances
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.12
texttable==1.6.7
typing_extensions==4.5.0
ulid-py==1.1.0
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.23.0
xlrd==2.0.1
xlwt==1.3.0
